Desktop: Update CSS variables in user iframes on theme change

+++ packages/app-desktop/services/plugins/hooks/useThemeCss.ts
@@ -36,16 +36,18 @@ export default function useThemeCss(dep: HookDependencies) {
 	const [cssFilePath, setCssFilePath] = useState('');
 
 	useEffect(() => {
-		if (cssFilePath) return () => {};
-
 		let cancelled = false;
 
 		async function createThemeStyleSheet() {
 			const theme = themeStyle(themeId);
 			const css = themeToCssVariables(theme);
 			const filePath = `${Setting.value('tempDir')}/plugin_${pluginId}_theme_${themeId}.css`;
-			await shim.fsDriver().writeFile(filePath, css, 'utf8');
-			if (cancelled) return;
+
+			if (!(await shim.fsDriver().exists(filePath))) {
+				await shim.fsDriver().writeFile(filePath, css, 'utf8');
+				if (cancelled) return;
+			}
+
 			setCssFilePath(filePath);
 		}
 
@@ -54,7 +56,7 @@ export default function useThemeCss(dep: HookDependencies) {
 		return () => {
 			cancelled = true;
 		};
-	}, [pluginId, themeId, cssFilePath]);
+	}, [pluginId, themeId]);
 
 	return cssFilePath;
 }
